商品数据控制开发

+++ datamodels/order.go
@@ -4,7 +4,7 @@ type Order struct {
 	ID          int64 `sql:"ID"`
 	UserId      int64 `sql:"userID"`
 	ProductId   int64 `sql:"productID"`
-	OrderStatus int64 `sql:"orderStatus"`
+	OrderStatus int   `sql:"orderStatus"`
 }
 
 const (
